Refactor(tests, #10066): consolidate notebook_tools shadow tests — tranche 1/3
Three legacy shadow files in scripts/tests/ consolidated into canonical

scripts/notebook_tools/tests/. Per CLAUDE.md consolidation protocol

(ANALYZE → MERGE → ARCHIVE):

- test_detect_ascii_workaround.py: 8 unique-to-legacy tests ported

  (TestSkipDirs 5 + TestIterNotebooksExcludesOutput 1 covering _should_skip

  + _iter_notebooks artifact filter contract; CSharpBar 8 already in canon)

- test_detect_blank_figures.py: clear superset (legacy 27 → canon 77),

  zero unique-to-legacy, ARCHIVE direct

- test_detect_svg_decimal_commas.py: 3 unique-to-legacy tests ported

  (test_graphviz_integer_coords_not_flagged for #6946 anti-FP regression,

  test_empty_points_not_flagged, test_detect_cell_handles_list_html_blob)

169/169 canonical tests PASS. Net -432 lines (-530 legacy, +98 new tests).

See #10066 (partial sub-grain 1/3 of 12 shadow test files).
